v2/requirements.txt
flask
pymongo==2.4.2
gunicorn
Flask-Assets==0.8
jsmin
cssmin
BeautifulSoup4
requests
redis
pypng
